Split the MQTT code, separate the transport link layer from the output driver
